+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 1.0.2
+
+- Rename a class from Request to Fetcher
+- Separate request and response classes for each mode
+- Add request jsonl output feature
+- Add batch request feature
+- Fix endpoint urls when location_id is global
+- Change required ruby version from 2.7 to 3.1
+
 ## 1.0.1
 
 - Add HTTP response code methods in Llmclt::Response